# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Skedway",
        "description": "Documentação Skedway.",
        "contact": {
            "email": "bernardodev0809@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/reservation/create": {
            "post": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Criar uma reserva",
                "operationId": "ea0a9c2ec34a65661203f4da2891426d",
                "requestBody": {
                    "description": "Dados da reserva",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roomId",
                                    "dateInit",
                                    "dateEnd"
                                ],
                                "properties": {
                                    "roomId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dateInit": {
                                        "type": "string",
                                        "example": "2025-08-29 21:58"
                                    },
                                    "dateEnd": {
                                        "type": "string",
                                        "example": "2025-08-29 22:58"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva criada!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "room_id": {
                                                    "type": "string",
                                                    "example": 1
                                                },
                                                "dateInit": {
                                                    "type": "string",
                                                    "example": "2025-08-29 22:01:00"
                                                },
                                                "dateEnd": {
                                                    "type": "string",
                                                    "example": "2025-08-29 23:01:00"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Período de reserva indisponível!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Sala reservada na data requerida. Por favor, escolha outra data!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reservation/cancel": {
            "patch": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Cancelar uma reserva",
                "operationId": "68682b728f57b25a4695cb067a09cbb2",
                "requestBody": {
                    "description": "Id da reserva",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva cancelada!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Reserva cancelada com sucesso!"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao cancelar reserva!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Reserva não pode ser cancelada, tente novamente mais tarde!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}